Support explicitly disabling interactivity & TTY flags for Docker in "deployster.sh".
Currently, "deployster.sh" sends the "--interactive" and "--tty" flags to "docker run" if it believes that the script is running in a terminal. However, since the test is not perfect, it makes sense to allow the caller to disable those flags explicitly.

Thus the caller can set the "TTY" environment variable to "false", "0", "off" or "no" and these flags WILL NOT be sent to "docker run".